# src/main/java/com/akiner/brokage_firm/rest/dto/StockOrderDto.java
package com.akiner.brokage_firm.rest.dto;

import com.akiner.brokage_firm.model.OrderSide;

public class StockOrderDto {

    private Long customerId;
    private String assetName;
    private OrderSide orderSide; // BUY or SELL
    private int size;
    private double price;

    public Long getCustomerId() {
        return customerId;
    }

    public void setCustomerId(Long customerId) {
        this.customerId = customerId;
    }

    public String getAssetName() {
        return assetName;
    }

    public void setAssetName(String assetName) {
        this.assetName = assetName;
    }

    public OrderSide getOrderSide() {
        return orderSide;
    }

    public void setOrderSide(OrderSide orderSide) {
        this.orderSide = orderSide;
    }

    public int getSize() {
        return size;
    }

    public void setSize(int size) {
        this.size = size;
    }

    public double getPrice() {
        return price;
    }

    public void setPrice(double price) {
        this.price = price;
    }
}